model/back_translation/requirements.txt
coverage==4.5.4
pytest==5.1.1
pytest-cov==2.7.1
python-dotenv==0.10.3
selenium==3.141.0
